Move sanitization logic to NullBlogPropertySanitizer
